Update cover.py to Support Opening/Closing/Stopped States with is_closing and is_opening
Hello,

Sorry Its probably not the way I should do that, but I am very new to Github and HomeAssistant, please forgive me :-)
Also I am not programmer so I would like to ask if someone can do a change for me. 
I'm also not sure if I will phrase that correctly. Sorry about that if its the case. I will try to provide the max info.

Is it possible for someone to add the support of additional states (OPENING and CLOSING and STOPPED) with the usage of an additional MQTT topic. I think this a supported feature of the platform/entity ? with is_opening or is_closing ? (https://developers.home-assistant.io/docs/en/entity_cover.html#supported-features)

There would be 3 different payloads which would define the "direction" of the shutter.

here is an example of my homebridge equivalent with tasmota shutter if it helps to understand : 

shutter2/tele/RESULT this topic from my tasmota provides this JSON message 
{
    "Shutter1": {
        "Position": 100,
        "direction": 0
    }
}

So, with direction being 1, 0 or -1 I can know if it goes up, down or if it is stopped

in Homebridge ( Homebridge can do that with "PostionStateValue" : https://www.npmjs.com/package/homebridge-mqttthing#window-covering ),
 this works :

 "getPositionState": {
                    "topic": "shutter2/tele/RESULT",
                    "apply": "return JSON.parse(message).Shutter1.direction;"
                }
            },
            "positionStateValues": [
                1,
                -1,
                0 "getPositionState": {
                    "topic": "shutter2/tele/RESULT",
                    "apply": "return JSON.parse(message).Shutter1.direction;"
                }
            },
            "positionStateValues": [
                1,
                -1,
                0
            ]

Thank you very much 👍